# cmd/server/openapi.yaml
openapi: 3.0.3
info:
  title: Datumpy IoT Platform API
  description: |
    High-performance IoT data ingestion and device management platform.
    
    ## Authentication
    - **User Auth**: JWT token in `Authorization: Bearer {token}` header
    - **Device Auth**: API key in `Authorization: Bearer {api_key}` header
    
    ## Quick Start
    1. Register user: `POST /auth/register`
    2. Create device: `POST /devices`
    3. Send data (HTTP): `POST /data/{device_id}`

    ## MQTT / IoT Integration
    For real-time data and low-overhead communication, use MQTT:
    - **Brokers**:
      - `TCP`: Port `1883` (No TLS)
      - `MQTTS`: Port `8883` (Secure TLS)
      - `WS`: Port `1884` (WebSocket Secure - via Traefik)
    - **Authentication**:
      - Username: `{device_id}`
      - Password: `{api_key}`
    - **Topics**:
      - Data (Pub): `data/{device_id}`
      - Commands (Sub): `commands/{device_id}`
      - Firmware (Sub): `firmware/{device_id}`
  version: 1.1.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://datum.bezg.in
    description: Production Server
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User Authentication. Accepts standard JWT Access Token (via Login) OR Persistent User API Key (starts with `ak_`).
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Device Authentication. Accepts Device API Key (starts with `sk_`) OR Dynamic Device Token (starts with `dk_`).
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    CreateKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user_id:
          type: string
...
paths:
...
  /auth/register:
    post:
      summary: Register new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: User already exists
  /auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/keys:
    post:
      summary: Create User API Key
      tags: [Authentication]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyRequest'
      responses:
        '201':
          description: Key created (Save the key!)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResponse'
    get:
      summary: List User API Keys
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyResponse'
  /auth/keys/{id}:
    delete:
      summary: Revoke User API Key
      tags: [Authentication]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Key deleted
  /auth/forgot-password:
    post:
      summary: Request password reset email
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Email sent (if user exists)
  /auth/reset-password:
    post:
      summary: Complete password reset
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated
  /auth/password:
    put:
      summary: Change password
      tags: [Authentication]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
  /auth/user:
    delete:
      summary: Delete account
      description: Permanently delete the authenticated user and all their devices/data.
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deleted
  /devices/firmware/{filename}:
    get:
      summary: Download Firmware (Secured)
      tags: [Devices]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Device API key
      responses:
        '200':
          description: Firmware binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /devices:
    post:
      summary: Create device
      tags: [Devices]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
    get:
      summary: List devices
      tags: [Devices]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of devices
  /devices/register:
    post:
      summary: Register device (Mobile App)
      tags: [Provisioning]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Device registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDeviceResponse'
...
  /devices/{device_id}:
    delete:
      summary: Delete device
      tags: [Devices]
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Device deleted
  /devices/{device_id}/commands:
    post:
      summary: Send command to device
      tags: [Commands]
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandRequest'
      responses:
        '201':
          description: Command queued
  /devices/{device_id}/commands/pending:
    get:
      summary: Poll for commands (device)
      tags: [Commands]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pending commands
  /devices/{device_id}/commands/poll:
    get:
      summary: Long poll for commands
      tags: [Commands]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
        - name: wait
          in: query
          schema:
            type: integer
            default: 30
            maximum: 60
          description: Seconds to wait for commands
      responses:
        '200':
          description: Commands or timeout
  /devices/{device_id}/commands/{command_id}/ack:
    post:
      summary: Acknowledge command execution
      tags: [Commands]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
        - name: command_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Command acknowledged
  /devices/{device_id}/push:
    get:
      summary: Push data via GET (for constrained devices)
      description: |
        Allows constrained IoT devices (like ESP8266, Arduino with limited HTTP libraries) 
        to submit sensor data via simple GET request with query parameters.
        This is an alternative to POST /data/{device_id} for devices that cannot easily 
        construct POST requests with JSON bodies.
      tags: [Data]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
          description: Device ID
        - name: key
          in: query
          required: true
          schema:
            type: string
          description: Device API key (sk_live_XXX format)
        - name: temp
          in: query
          schema:
            type: number
          description: Temperature (example sensor value)
        - name: humidity
          in: query
          schema:
            type: number
          description: Humidity (example sensor value)
      responses:
        '200':
          description: Data stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                  fields_stored:
                    type: integer
                    description: Number of sensor values stored
                  commands_pending:
                    type: integer
  /devices/{device_id}/stream/frame:
    post:
      summary: Upload video frame
      tags: [Streaming]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Frame processed
  /devices/{device_id}/stream/mjpeg:
    get:
      summary: MJPEG Stream
      tags: [Streaming]
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MJPEG Stream
  /devices/{device_id}/stream/snapshot:
    get:
      summary: Get current frame snapshot
      tags: [Streaming]
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: JPEG Image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
  /devices/{device_id}/stream/ws:
    get:
      summary: WebSocket Stream
      tags: [Streaming]
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '101':
          description: WebSocket upgrade
  /devices/{device_id}/stream/info:
    get:
      summary: Stream Metadata
      tags: [Streaming]
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stream info
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                  fps:
                    type: number
                  width:
                    type: integer
                  height:
                    type: integer
  /data/{device_id}:
    post:
      summary: Send data (device)
      tags: [Data]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataPayload'
      responses:
        '200':
          description: Data received
    get:
      summary: Get latest data
      tags: [Data]
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Latest data point
...
  /public/data/{device_id}:
    post:
      summary: Send public data (no auth)
      tags: [Public]
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataPayload'
      responses:
        '200':
          description: Data received
    get:
      summary: Get public data (no auth)
      tags: [Public]
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Latest data
  /public/data/{device_id}/history:
    get:
      summary: Get public history with time range filtering (no auth)
      description: |
        Query historical time-series data without authentication.
        Same time range and aggregation options as authenticated endpoint.
      tags: [Public]
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of data points
          schema:
            type: integer
            default: 100
        - name: range
          in: query
          description: Duration backwards from now (e.g., 1h, 24h, 7d)
          schema:
            type: string
        - name: start
          in: query
          description: Start time as Unix timestamp in milliseconds
          schema:
            type: integer
            format: int64
        - name: stop
          in: query
          description: Duration from start (e.g., 1h, 1d)
          schema:
            type: string
        - name: int
          in: query
          description: Aggregation interval (e.g., 1m, 1h, 1d)
          schema:
            type: string
      responses:
        '200':
          description: Historical data points
  /devices/{device_id}/commands/stream:
    get:
      summary: SSE stream for real-time commands
      description: |
        Server-Sent Events (SSE) endpoint for real-time command delivery.
        Keeps connection open and pushes commands as they arrive.
        Battery-optimized for IoT devices.
        
        **Event Types:**
        - `command`: New command received
        - `ping`: Keep-alive (every 30s)
      tags: [Commands]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: command
                data: {"id":"cmd123","action":"reboot","params":{}}
                
                event: ping
                data: {"time":"2024-01-01T12:00:00Z"}


  # ============ Admin Endpoints ============
...